requirements.txt
fastapi==0.116.1
requests==2.32.4
python-dotenv==1.1.1
psycopg2==2.9.10